v1.3.8: Refactor Phase 5 - OptionsPanel Decomposition
