+++ src/components/options/index.ts
@@ -1,0 +1,4 @@
+export { TimetableManager } from './TimetableManager';
+export { BackgroundSettings } from './BackgroundSettings';
+export { AppSettings } from './AppSettings';
+export { PrivacySection } from './PrivacySection';
